Initial commit
This started as me just fucking about in flash for half an hour,
but by the end i had something i wanted to keep developing.
As a result it is written unmaintainably and without source
control. I am commiting all my current code and will be improving
it steadily in small frequent commits as you are supposed to.